refactor　mouse以外はできた
